+++ src/index.ts
@@ -21,4 +21,4 @@ export {
   PermissionDeniedError,
   UnprocessableEntityError,
 } from './core/error';
-export { TelnyxWebhook, TelnyxWebhookVerificationError } from './webhooks';
+export { TelnyxWebhook, TelnyxWebhookVerificationError } from './lib/webhooks';