+++ src/resources/verifications/index.ts
@@ -10,4 +10,5 @@ export {
   type VerificationTriggerCallParams,
   type VerificationTriggerFlashcallParams,
   type VerificationTriggerSMSParams,
+  type VerificationTriggerWhatsappVerificationParams,
 } from './verifications';